refactor: renamed the constant file to helper
